refactor: opt 拆出 metrics 模块;删掉无人调用的 OptStats::is_empty
opt.rs 里混着两类性质相反的代码:会改 MIR 的 pass,和只观察不改的计数器。
对一个内存安全敏感的模块,'哪些代码会动 MIR'应该一眼看得出来。
拆出 opt/metrics.rs(826 → 611 + 224),模块头写明:这里没有任何东西
mutate MirModule,存在的意义是让'该不该写这个 pass'有数据支撑——其中
两个已经分别劝退了 LICM 和跨块 scope drop。

OptStats::is_empty 是我加的但全库无人调用的死代码,删。

workspace 110 全绿;clippy 零警告;coverage 51/51;stats 输出不变
